requirements.txt
streamlit
python-dotenv
streamlit-audiorecorder
azure-cognitiveservices-speech
gspread
oauth2client
pydub
